+++ SpringMvc/src/main/java/com/victoria/model/Order.java
@@ -1,20 +1,20 @@
 package com.victoria.model;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public class Order {
     private BigInteger orderId;
     private BigInteger userId;
-    private BigDecimal orderCost;
+    private BigInteger orderCost;
     private String status;
     private BigInteger courierId;
-    public Order(BigInteger orderId, BigInteger userId, BigDecimal orderCost, String status, BigInteger courierId) {
+
+    public Order(BigInteger orderId, BigInteger userId, BigInteger orderCost, String status, BigInteger courierId) {
         this(orderId, userId, orderCost, status);
         this.courierId = courierId;
     }
 
-    public Order(BigInteger orderId, BigInteger userId, BigDecimal orderCost, String status) {
+    public Order(BigInteger orderId, BigInteger userId, BigInteger orderCost, String status) {
         this.orderId = orderId;
         this.userId = userId;
         this.orderCost = orderCost;
@@ -33,7 +33,7 @@ public class Order {
         return userId;
     }
 
-    public BigDecimal getOrderCost() {
+    public BigInteger getOrderCost() {
         return orderCost;
     }
 
@@ -52,7 +52,7 @@ public class Order {
     public void setCourierId(BigInteger userId) {
         this.courierId = userId;
     }
-    public void setOrderCost(BigDecimal orderCost) {
+    public void setOrderCost(BigInteger orderCost) {
         this.orderCost = orderCost;
     }
